Reorganise JAX/EasyDel helpers under tools/jax/{,easydel/}
Move JAX-only helpers into a dedicated tools/jax/ package and EasyDel-specific
helpers into tools/jax/easydel/. Drop the redundant jax_/easydel_ filename
prefixes; symbol names (JaxDeviceManager, JaxCheckpointManager, ...) are kept
unchanged so the diff stays mechanical.

  tools/jax_helpers.py             -> tools/jax/helpers.py
  tools/jax_device_manager.py      -> tools/jax/device_manager.py
  tools/jax_checkpoint_manager.py  -> tools/jax/checkpoint_manager.py
  tools/easydel_helpers.py         -> tools/jax/easydel/helpers.py
  tools/easydel_partitioning.py    -> tools/jax/easydel/partitioning.py
  tools/easydel_train_steps.py     -> tools/jax/easydel/train_steps.py
  tools/workaround_utils_jax.py    -> tools/jax/easydel/workaround_utils.py

Torch-specific and shared (logging_manager, reproducibility_manager, cli,
storage_backends, templates, ...) helpers stay at tools/ and will be
restructured in follow-up PRs.

Cross-imports inside the moved files plus all callers in
experiments/easydel/qwen and experiments/jax/distil_bert have been updated.
A single 'import easydel.modules' bootstrap is added to
tools/jax/easydel/workaround_utils.py to remove a latent dependency on
import order with EasyDel's lazy submodule loading.
