+++ blacksmith/experiments/jax/distil_bert/multi_chip/data_parallel/train.py
@@ -25,7 +25,7 @@ from blacksmith.experiments.jax.distil_bert.multi_chip.data_parallel.sharding_co
 from blacksmith.models.jax.distil_bert.model import init_model
 from blacksmith.models.jax.distil_bert.model_utils import combine_params, split_params
 from blacksmith.tools.cli import generate_config, parse_cli_options
-from blacksmith.tools.jax_helpers import (
+from blacksmith.tools.jax.helpers import (
     build_schedule,
     ce_with_labels,
     cosine_embedding_loss,
